1srpbothrdaily/pyproject.toml
[project]
name = "5rsimaalgotrading"
version = "0.1.0"
description = "Daily Breakout Trading Bot for Delta Exchange"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "delta-rest-client>=1.0.13",
    "python-dotenv>=1.0.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "pyyaml>=6.0",
    "pytz>=2024.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
